Create data model for prototype.
This commit creates a number of model classes, each of which is used to model
data arising from a metric used within the prototype. The model classes are
missing a lot of functionality which will be required.

Release models a planned release of a product. This will involve a number of
user stories comprising the features for this release. The release itself
will involve a number of sprints, each of which relates to one team.

The team (as already mentioned) undertakes sprints over the course of
development. Periodically, defects will be discovered and assigned to a team
for resolution. Additionally, teams will measure happiness & satisfaction at
each retrospective, together with assessing their maturity in practice use.

All of this is modelled by the classes comprising this commit.